AppEventProcessing/requirements.txt
pyspark
black
pyyaml